📁 Reorganize themes into site and application directories
- Move lib/themes → themes/site (website themes)
- Move styles → themes/application (UI pattern themes)
- Update all imports to use new paths
- Add comprehensive README documentation explaining the separation

Site themes: Simple color/font variations for the website
Application themes: Complete UI theming system for apps (buttons, forms, states, etc.)
