+++ themes/application/README.md
@@ -1,0 +1,36 @@
+# Application Themes
+
+Comprehensive theming system for building consistent, professional application UIs. These themes define complete visual languages for interactive applications.
+
+## Purpose
+
+Unlike simple website themes, application themes provide:
+- Complete component styling (buttons, forms, cards, etc.)
+- Interactive states (hover, focus, active, disabled)
+- Data visualization colors and patterns
+- Consistent spacing and sizing systems
+- Accessibility considerations
+
+## Available Themes
+
+- **Terminal** - Classic terminal/CLI aesthetic
+- **Terminal Hybrid** - Terminal with CSS variable enhancements
+- **Directory** - File explorer inspired theme
+
+## Theme Capabilities
+
+Application themes define styling for:
+- **Forms**: Inputs, selects, checkboxes, radios
+- **Buttons**: Primary, secondary, ghost, sizes, states
+- **Feedback**: Alerts, toasts, notifications
+- **Layout**: Cards, panels, sidebars
+- **Data**: Tables, charts, graphs
+- **Navigation**: Menus, tabs, breadcrumbs
+
+## Usage
+
+These themes are primarily used by the styleguide (`/styleguide`) to demonstrate the design system capabilities. They serve as a reference implementation for building cohesive application interfaces.
+
+## Adding Application Themes
+
+New themes should provide comprehensive coverage of all UI patterns. See existing themes for the expected structure and scope.
