+++ themes/application/terminal-theme-hybrid.ts
