+++ themes/application/terminal-theme.ts
