+++ themes/application/theme-registry.ts
